Extract class Partition from PartitionTable
As the first step in providing table statistics, this commit
extracts class Partition from PartitionTable, and passes the
two PartitionTable attributes referenced by Partition members
as constructor parameters.
